Make Launch services an independent function.
Since the launcher is used one time at the beginning and never
afterwards during the steady state, there is no need to make this a
separate type managed by service manager.

Internally this function still uses a type along with functions for
convenience, but it is still a fire and forget model.